json locales and littel change

+++ server/server.lua
@@ -1,6 +1,7 @@
 local Config = lib.require('config')
+lib.locale()
 
-RSGCore.Commands.Add('myjobs', 'Opens your job menu', {}, false, function(source)
+RSGCore.Commands.Add('myjobs', locale('sv_command_desc'), {}, false, function(source)
     local src = source
     local Player = RSGCore.Functions.GetPlayer(src)
     TriggerClientEvent('rsg-multijob:client:openmenu', src)
@@ -27,13 +28,14 @@ end
 
 local function CanSetJob(cid, jobName)
     local jobs = MySQL.query.await('SELECT job, grade FROM player_jobs WHERE citizenid = ? ', {cid})
-    if not jobs then return false end
-    for i = 1, #jobs do
-        if jobs[i].job == jobName then
-            return true, jobs[i].grade
+    if not jobs then return false, nil end
+
+    for _, jobData in ipairs(jobs) do
+        if jobData.job == jobName then
+            return true, jobData.grade
         end
     end
-    return false
+    return false, nil
 end
 
 lib.callback.register('rsg-multijob:server:myJobs', function(source)
@@ -44,14 +46,14 @@ lib.callback.register('rsg-multijob:server:myJobs', function(source)
     for k, v in pairs(result) do
         local job = RSGCore.Shared.Jobs[v.job]
 
-        if not job then 
-            return error(('MISSING JOB FROM jobs.lua: "%s" | CITIZEN ID: %s'): format(v.job, Player.PlayerData.citizenid)) 
+        if not job then
+            return error(('MISSING JOB FROM jobs.lua: "%s" | CITIZEN ID: %s'): format(v.job, Player.PlayerData.citizenid))
         end
-        
+
         local grade = job.grades[tostring(v.grade)]
 
-        if not grade then 
-            return error(('MISSING JOB GRADE for "%s". GRADE MISSING: %s | CITIZEN ID: %s'): format(v.job, v.grade, Player.PlayerData.citizenid)) 
+        if not grade then
+            return error(('MISSING JOB GRADE for "%s". GRADE MISSING: %s | CITIZEN ID: %s'): format(v.job, v.grade, Player.PlayerData.citizenid))
         end
 
         storeJobs[#storeJobs + 1] = {
@@ -69,28 +71,28 @@ RegisterNetEvent('rsg-multijob:server:changeJob', function(job)
     local src = source
     local Player = RSGCore.Functions.GetPlayer(src)
 
-    if Player.PlayerData.job.name == job then 
-        TriggerClientEvent('ox_lib:notify', src, {title = 'Your current job is already set to this.', type = 'error', duration = 5000 })
+    if Player.PlayerData.job.name == job then
+        TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_current_job_error'), type = 'error', duration = 5000 })
         return 
     end
 
     local jobInfo = RSGCore.Shared.Jobs[job]
-    if not jobInfo then 
-        TriggerClientEvent('ox_lib:notify', src, {title = 'Invalid job.', type = 'error', duration = 5000 })
-        return 
+    if not jobInfo then
+        TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_invalid_job'), type = 'error', duration = 5000 })
+        return
     end
 
     local cid = Player.PlayerData.citizenid
     local canSet, grade = CanSetJob(cid, job)
-    
-    if not canSet then 
-        return 
+
+    if not canSet then
+        return
     end
 
     Player.Functions.SetJob(job, grade)
     Player.Functions.SetJobDuty(false)
     TriggerClientEvent('RSGCore:Client:SetDuty', src, false)
-    TriggerClientEvent('ox_lib:notify', src, {title = 'Your job is now: ' .. jobInfo.label, type = 'info', duration = 5000 })
+    TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_job') .. ': ' .. jobInfo.label, type = 'info', duration = 5000 })
 end)
 
 RegisterNetEvent('rsg-multijob:server:newJob', function(newJob)
@@ -105,18 +107,19 @@ RegisterNetEvent('rsg-multijob:server:newJob', function(newJob)
         hasJob = true
         return
     end
-    if not hasJob and GetJobCount(cid) < Config.MaxJobs then 
-        MySQL.insert.await('INSERT INTO player_jobs (citizenid, job, grade) VALUE (?, ?, ?)', {cid, newJob.name, newJob.grade.level})
-    else
-        return TriggerClientEvent('ox_lib:notify', src, {title = 'You have the max amount of jobs.', type = 'error', duration = 5000 })
+    if GetJobCount(cid) >= Config.MaxJobs then
+        return TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_job_max'), type = 'error', duration = 5000 })
     end
+    -- if not hasJob then
+    MySQL.insert.await('INSERT INTO player_jobs (citizenid, job, grade) VALUE (?, ?, ?)', {cid, newJob.name, newJob.grade.level})
+    -- end
 end)
 
 RegisterNetEvent('rsg-multijob:server:deleteJob', function(job)
     local src = source
     local Player = RSGCore.Functions.GetPlayer(src)
     MySQL.query.await('DELETE FROM player_jobs WHERE citizenid = ? and job = ?', {Player.PlayerData.citizenid, job})
-    TriggerClientEvent('ox_lib:notify', src, {title = 'You deleted '..RSGCore.Shared.Jobs[job].label..' job from your menu.', type = 'success', duration = 5000 })
+    TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_job_deleted') .. ' ' .. RSGCore.Shared.Jobs[job].label .. ' '.. locale('sv_job_deleted_2'), type = 'success', duration = 5000 })
     if Player.PlayerData.job.name == job then
         Player.Functions.SetJob('unemployed', 0)
     end
@@ -144,6 +147,9 @@ local function adminRemoveJob(src, id, job)
     local Player = RSGCore.Functions.GetPlayer(id)
     local cid = Player.PlayerData.citizenid
     local result = MySQL.query.await('SELECT * FROM player_jobs WHERE citizenid = ? AND job = ?', {cid, job})
+    if not result[1] then
+        return TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_job_specified'), type = 'error', duration = 5000 })
+    end
     if result[1] then
         MySQL.query.await('DELETE FROM player_jobs WHERE citizenid = ? AND job = ?', {cid, job})
         TriggerClientEvent('ox_lib:notify', src, {title = ('Job: %s was removed from ID: %s'):format(job, id), type = 'success', duration = 5000 })
@@ -151,25 +157,23 @@ local function adminRemoveJob(src, id, job)
             Player.Functions.SetJob('unemployed', 0)
         end
     else
-        TriggerClientEvent('ox_lib:notify', src, {title = 'Player doesn\'t have this job?', type = 'error', duration = 5000 })
+        TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_job_does'), type = 'error', duration = 5000 })
     end
 end
 
-RSGCore.Commands.Add('removejob', "Remove a job from the player's multijob.", { { name = 'id', help = 'ID of the player' }, { name = 'job', help = 'Name of Job' } }, true, function(source, args)
+RSGCore.Commands.Add('removejob', locale('sv_command_remove'), { { name = 'id', help = locale('sv_command_r_id')  }, { name = 'job', help = locale('sv_command_r_name') } }, true, function(source, args)
     local src = source
-    if not args[1] then 
-        TriggerClientEvent('ox_lib:notify', src, {title = 'Must provide a player id.', type = 'error', duration = 5000 })
+    if not args[1] then
+        TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_provide'), type = 'error', duration = 5000 })
         return
     end
-    if not args[2] then 
-        TriggerClientEvent('ox_lib:notify', src, {title = 'Must provide the name of the job to remove from the player.', type = 'error', duration = 5000 })
-        return 
+    if not args[2] then
+        TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_provide_name'), type = 'error', duration = 5000 })
+        return
     end
     local id = tonumber(args[1])
     local Player = RSGCore.Functions.GetPlayer(id)
-    if not Player then 
-        TriggerClientEvent('ox_lib:notify', src, {title = 'Player not online.', type = 'error', duration = 5000 })
-        return
-    end
+    if not Player then TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_not_online'), type = 'error', duration = 5000 }) return end
+
     adminRemoveJob(src, id, args[2])
 end, 'admin')
